Initial Commit
Simulator program for Control Systems using a Block oriented approach.
This was an initial inception by myself in college in 1995. I used it
to prove that we could develop such a system from scratch avoiding
acquisition of expensive software. Unfortunately my vision was not
shared, and by lack of my diplomatic skills probably seen as a direct
attack on the virtues of the docent I was trying to convince … I think
he didn’t like me …